+++ rules/github.cjs
@@ -1,0 +1,18 @@
+// https://github.com/github/eslint-plugin-github#rules
+
+/** @type {import("eslint").Linter.Config["rules"]} */
+module.exports = {
+  "github/array-foreach": "error",
+  "github/async-currenttarget": "error",
+  "github/async-preventdefault": "error",
+  "github/get-attribute": "error",
+  "github/no-blur": "error",
+  "github/no-dynamic-script-tag": "error",
+  "github/no-implicit-buggy-globals": "error",
+  "github/no-inner-html": "error",
+  "github/no-then": "error",
+  "github/no-useless-passive": "error",
+  "github/prefer-observers": "error",
+  "github/require-passive-events": "error",
+  "github/unescaped-html-literal": "error"
+};