+++ test/rando.ts
@@ -1,1 +1,25 @@
 console.log(undefined!);
+
+const strins = ['a', 'b', 'c'];
+
+if(!false) console.log('sug');
+if(!false) console.log('sug');
+
+var a = 1;
+let b =- a
+
+async function x(y: string): Promise<number> {
+    return 1;
+}
+
+x('', 5).then(() => {});
+
+
+
+'hello'.lastIndexOf("e");
+
+strins.forEach(s => {
+  const el = document.querySelector(s) as HTMLElement;
+  el.innerHTML = 'hello';
+  el.getAttribute('aria-Lable');
+})